backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
PyNaCl==1.5.0
base58==2.1.1
